requirements.txt
numpy
PyOpenGL
glfw
pillow
pyrr
pygame